adk_agent/requirements.txt
google-adk==1.3.0
fastapi
uvicorn
pg8000
python-dotenv
opentelemetry-sdk
google-cloud-logging

